+++ consul/aio.py
@@ -19,12 +19,12 @@ class HTTPClient(base.HTTPClient):
         super(HTTPClient, self).__init__(*args, **kwargs)
         self._loop = loop or asyncio.get_event_loop()
 
-    async def _request(self, callback, method, uri, data=None):
+    async def _request(self, callback, method, uri, data=None, headers=None):
         connector = aiohttp.TCPConnector(loop=self._loop,
                                          verify_ssl=self.verify)
         async with aiohttp.ClientSession(connector=connector) as session:
             self._session = session
-            resp = await session.request(method=method, url=uri, data=data)
+            resp = await session.request(method=method, url=uri, data=data, headers=headers)
             body = await resp.text(encoding='utf-8')
             content = await resp.read()
             if resp.status == 599:
@@ -43,21 +43,21 @@ class HTTPClient(base.HTTPClient):
             #                   ResourceWarning)
             # self._session.close()
 
-    async def get(self, callback, path, params=None):
+    async def get(self, callback, path, params=None, headers=None):
         uri = self.uri(path, params)
-        return await self._request(callback, 'GET', uri)
+        return await self._request(callback, 'GET', uri, headers=headers)
 
-    async def put(self, callback, path, params=None, data=''):
+    async def put(self, callback, path, params=None, data='', headers=None):
         uri = self.uri(path, params)
-        return await self._request(callback, 'PUT', uri, data=data)
+        return await self._request(callback, 'PUT', uri, data=data, headers=headers)
 
-    async def delete(self, callback, path, params=None, data=''):
+    async def delete(self, callback, path, params=None, data='', headers=None):
         uri = self.uri(path, params)
-        return await self._request(callback, 'DELETE', uri, data=data)
+        return await self._request(callback, 'DELETE', uri, data=data, headers=headers)
 
-    async def post(self, callback, path, params=None, data=''):
+    async def post(self, callback, path, params=None, data='', headers=None):
         uri = self.uri(path, params)
-        return await self._request(callback, 'POST', uri, data=data)
+        return await self._request(callback, 'POST', uri, data=data, headers=headers)
 
     # async def close(self):
     #     await self._session.close()